requirements.txt
requests
pandas
matplotlib
seaborn
google-generativeai
python-dotenv